moved test classes around
