+++ Src/Newtonsoft.Json.Tests/Serialization/NewPropertyTests.cs
@@ -2,47 +2,26 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using Newtonsoft.Json.Tests.TestObjects;
 using NUnit.Framework;
 
 namespace Newtonsoft.Json.Tests.Serialization
 {
-    public class DerivedWithNewProperty<T>: SimpleBaseClass
-    {
-        public new T Data { get; set; }
-    }
-
-    public abstract class SimpleBaseClass
-    {
-        public string Result { get; set; }
-        public string Message { get; set; }
-        public object Data { get; set; }
-
-        protected SimpleBaseClass()
-        {
-            
-        }
-
-        protected SimpleBaseClass(string message)
-        {
-            Message = message;
-        }
-    }
-
     [TestFixture]
     public class NewGenericPropertyOnADerivedClassTests: TestFixtureBase
     {
         [Test]
         public void CanSerializeWithBuiltInTypeAsGenericArgument()
         {
-            var input = new DerivedWithNewProperty<int>()
+            var input = new ResponseWithNewGenericProperty<int>()
             {
-                Message = "Something",
+                Message = "Trying out integer as type parameter",
                 Data = 25,
                 Result = "This should be fine"
             };
 
             var json = JsonConvert.SerializeObject(input);
-            var deserialized = JsonConvert.DeserializeObject<DerivedWithNewProperty<int>>(json);
+            var deserialized = JsonConvert.DeserializeObject<ResponseWithNewGenericProperty<int>>(json);
 
             Assert.AreEqual(input.Data, deserialized.Data);
             Assert.AreEqual(input.Message, deserialized.Message);
@@ -52,15 +31,15 @@ namespace Newtonsoft.Json.Tests.Serialization
         [Test]
         public void CanSerializedWithGenericClosedTypeAsArgument()
         {
-            var input = new DerivedWithNewProperty<List<int>>()
+            var input = new ResponseWithNewGenericProperty<List<int>>()
             {
-                Message = "Something",
+                Message = "More complex case - generic list of int",
                 Data = Enumerable.Range(50, 70).ToList(),
-                Result = "This should be fine"
+                Result = "This should be fine too"
             };
 
             var json = JsonConvert.SerializeObject(input);
-            var deserialized = JsonConvert.DeserializeObject<DerivedWithNewProperty<List<int>>>(json);
+            var deserialized = JsonConvert.DeserializeObject<ResponseWithNewGenericProperty<List<int>>>(json);
 
             Assert.AreEqual(input.Data, deserialized.Data);
             Assert.AreEqual(input.Message, deserialized.Message);
